+++ apps/web/src/widgets/bottomSheet/index.ts
@@ -1,0 +1,1 @@
+export { BottomSheet } from "./ui/BottomSheet";
