+++ apps/web/src/widgets/bottomSheet/ui/BottomSheet.tsx
