+++ apps/web/src/widgets/tabs/index.ts
@@ -1,0 +1,1 @@
+export { Tabs } from "./ui/Tabs";
