+++ apps/web/src/widgets/tabs/ui/Tabs.tsx
